[refactor] fetch language fns to a separate module
